Add a FIFO class and use it to replace the code in bytestreamToUMP.
Replace the readIndex, writeIndex, bufferLength, and umpMessage fields
with a single FIFO instance. The separates the responsibility of
maintaining the FIFO and simplifies the more interesting code!

This commit only affects the FIFO in this one header file. There are
others that There are other ad-hoc FIFOs in the code base that can
also perhaps be switched over. I want to limit the scope of this
change for the time being.

That only leaves the quesiton of how to unit-test the code...