Add a using/typedef for each std::function<> to avoid duplication.
Previously the full signature for each of the UMP processor callbacks
was present twice in the code: once in the definition of the member
variable and again in the signature of the function that sets the
member variable. This refactor simply adds a typedef of each
std::function<> declaration and uses it in both of those places.